// src/model/AttendanceStatistics.ts
export interface AttendanceStatistics {
  present: number;
  absent: number;
  notChosen: number;
  poissonCount: number;
  viandeCount: number;
  enfantCount: number;
  accompanyingGuestCount: number;
  adultCount: number;
  mainGuestCount: number;
  totalPeople: number;
  menuIncomplet: number;
}